Support TLS server and client certificate authentication.
All RPCs going back and forth between the different components have been
unauthenticated so far. In order to be able to use red-cloud
productively, there needs to be some form of authentication. TLS is a
simple one, so that's what I'm going for here.

The same certificate and key are used to talk to etcd when enabled. This
also means that TLS can only be enabled "wholesale", i.e. for red-cloud
and etcd. If there is ever a need to use separate TLS configs for the
two, it can still be implemented.