Add tracing and export monitoring data to Prometheus from the caretaker.
This will turn the caretaker into more or less a whitebox from the
operational perspective. It will be much simpler to determine what it is
doing, why operations failed or took a very long time, or to find any
bottlenecks. It also allows users to monitor or alert on error
rates/ratios.