Add a heartbeat mechanism to regularly refresh the list of nodes.
This will regularly (every 5 minutes; should eventually make this a
flag) poll etcd for nodes that are not yet registered, and register
them.

This is a bit of a workaround for nodes that fall through the cracks
e.g. because we fail to connect to them and then forget to try again.
Eventually we should find a better solution, but this will do for now